rename object input and output
